% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allmetrics.R
\name{allmetrics}
\alias{allmetrics}
\title{A dataframe of a set of metrics, given grouping variables}
\usage{
allmetrics(datin, ...,  na.rm = TRUE, sdat, survdat, fdat, cryptdat, invertclasses)
}
\arguments{
\item{datin}{\code{tbl_df} the main data-frame to use}

\item{...}{optional grouping variables.}

\item{sdat}{\code{tbl_df} giving site level data}

\item{survdat}{\code{tbl_df} giving survey level data}

\item{fdat}{\code{tbl_df} giving species trait data}

\item{cryptdat}{\code{tbl_df} giving cryptic fish families to exclude in Method 1 calculations}

\item{invertclasses}{\code{character} invert classes to include in calculations}
}
\value{
xout a \code{tbl_df} with the (optional) grouping
variables and a set of metric values.
}
\description{
A dataframe of a set of metrics, given grouping variables
}
\details{
Common metrics are calculated that include:
Method 1 fish richness (\code{m1_alpha})
Method 1 fish biomass (\code{m1_biomass})
Biomass of fish >= 20cm (method 1) (\code{B20})
The above three metrics exclude fish from families \code{cryptdat}
Method 2 fish richness (\code{m2_fish_alpha})
Method 2 invert richness, for select Classes (\code{m2_invert_alpha})
See \code{data(invertclasses)} for Class names
Functional group richness  (\code{FG_richness})
See \code{data(fdat)} and field \code{FuncGroup} for groups
Community temperature indext \code{CTI}
Method 2 urchin abundance, excluding genus Echinostrephus \code{urchin_abundance}

All values are calculated at survey level then
averages are taken at higher levels, if requested.
}
\examples{

}
\author{
Christopher J. Brown
}
