% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcmetrics.R
\name{calcmetrics}
\alias{calcmetrics}
\title{Calculate metrics for a sampling unit from survey level data}
\usage{
calcmetrics(datin, ..., sizecat = 20,
na.rm = T,
metrics = c( "alpha", "biomass", "bioatsize"))
}
\arguments{
\item{datin}{\code{tbl_df} the main data-frame to use}

\item{...}{optional grouping variables.}

\item{sizecat}{\code{numeric} giving size minimum size category to include in
calculation of \code{bioatsize}}

\item{metrics}{\code{character} vector of metrics to calculate. Default is
all of them.}
}
\value{
xout a \code{tbl_df} with the (optional) grouping
variables and the metric value(s)
}
\description{
Calculate metrics for a sampling unit from survey level data
}
\details{
Metrics are calculated per SurveyID, then the mean is
taken across grouping variables.
Metric alpha is number of species on a survey, metric biomass is biomass of
metric bioatsize is biomass of fish >= \code{sizecat}
}
\examples{

data(rls)
calcmetrics(rls, SurveyID, metrics = "alpha") # Alpha richness by survey
calcmetrics(rls, SurveyID, Method ) #All metrics averaged by survey and method.
}
\author{
Christopher J. Brown
}
