% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commtempindex.R
\name{commtempindex}
\alias{commtempindex}
\title{Calculate the community temperature index}
\usage{
commtempindex(datin, ...,
, tempvar = "ThermalMP_5_95", abundvar = "BioMass",
na.rm = T, dolog = TRUE)
}
\arguments{
\item{datin}{\code{tbl_df} A dataframe of species observations with a
temperature index for each observation.}

\item{...}{optional grouping variables.}

\item{tempvar}{\code{character} temperature index variable}

\item{abundvar}{\code{character} abundance/biomass variable to use}

\item{na.rm}{\code{logical} remove NAs?}

\item{log}{\code{logical} log response?}
}
\value{
xout a \code{tbl_df} with the (optional) grouping
variables and the CTI values.
}
\description{
Calculate the community temperature index
}
\details{
The CTI is calculated per SurveyID, then the mean is
taken across grouping variables.
The CTI is calculated: \eqn{\frac{\sum{T_j B_{i,j}}}{\sum{B_{i,j}}}}
Where there are n surveys i, \eqn{T_j} are the species thermal indexes
(e.g. thermal midpoint from a species range) and \eqn{B_{i,j}} are the
biomasses or abundances of each species at each site.
The default is the log of the abundance.
}
\examples{
data(rls)
data(fdat)
flim <- fdat \%>\% select(SpeciesID, Class, Family, ThermalMP_5_95)

r2 <- rls \%>\% filter(Method ==1) \%>\%
   inner_join(survdat) \%>\%
   inner_join(sdat) \%>\%
   inner_join(flim) \%>\%
   filter(Class \%in\% fish_classes())
xout <- r2 \%>\% commtempindex(Location, Ecoregion, tempvar = "ThermalMP_5_95", abundvar = "BioMass", dolog = TRUE)
hist(xout$CTI)
}
\author{
Christopher J. Brown
}
