% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cryptdat}
\alias{cryptdat}
\title{Cryptic fish families}
\format{data.frame:
\describe{
  \item{Family}{Cryptic fish families}
}}
\usage{
cryptdat
}
\description{
It is useful to filter them out sometimes.
}
\keyword{datasets}
