% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fdat}
\alias{fdat}
\title{Species level data}
\format{data.frame:
\describe{
  \item{SpeciesID}{Species ID code}
  \item{TAXONOMIC_NAME}{Latin binomial}
  \item{and so on}{many other variables...}
}}
\usage{
fdat
}
\description{
Traits etc
}
\keyword{datasets}
