% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_rich.R
\name{func_rich}
\alias{func_rich}
\title{Calculate functional group richness}
\usage{
func_rich(datin, ..., group = "SpeciesID",
na.rm = T)
}
\arguments{
\item{datin}{\code{tbl_df} the main data-frame to use}

\item{...}{optional grouping variables.}

\item{group}{\code{character} giving category to calculate
richness over}
}
\value{
xout a \code{tbl_df} with the (optional) grouping
variables and the richness value(s)
}
\description{
Calculate functional group richness
}
\details{
Richness is calculated per SurveyID, then the mean is
taken across grouping variables.
}
\examples{

}
\author{
Christopher J. Brown
}
