% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{invertclasses}
\alias{invertclasses}
\title{Invertebrate classes}
\format{character
It is useful to filter them out sometimes.}
\usage{
invertclasses
}
\description{
Invertebrate classes
}
\keyword{datasets}
