% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_classes.R, R/nspp.R
\name{fish_classes}
\alias{fish_classes}
\alias{nspp}
\title{Class names for fish and sharks for filtering data frames}
\usage{
fish_classes()

nspp(datin, ...)
}
\arguments{
\item{datin}{\code{tbl_df} the main data-frame to use
... optional grouping variables.}
}
\value{
xout a \code{character} vector
\code{c("Actinopterygii", "Elasmobranchii")}

xout a \code{tbl_df} with the (optional) grouping
variables and species richness.
}
\description{
Class names for fish and sharks for filtering data frames

Calculate gamma diversity for a sampling unit
}
\details{


Calculates total numbers of species across the groups i.e. gamma diversity
See \code{\link{calcmetrics}} to get mean survey level richness.
}
\examples{
data(fdat)
filter(fdat, Class \%in\% fish_classes())
nspp(rls, SurveyID) # Richness by survey
 nspp(rls, SurveyID, Method ) #Richness by survey and method.
}
\author{
Christopher J. Brown

Christopher J. Brown
}
