% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rls}
\alias{rls}
\title{Full RLS transect dataset}
\format{data.frame:
\describe{
  \item{SurveyID}{Survey ID code}
  \item{SpeciesID}{Species ID code}
  \item{SizeClass}{Size class in cm}
  \item{Method}{Which transect method in 0-3}
  \item{Abundance}{Count of observed individuals}
  \item{BioMass}{Weight in grams if calculated}
}}
\usage{
rls
}
\description{
Observations of species across global sites
}
\keyword{datasets}
