% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sdat}
\alias{sdat}
\title{Site level data}
\format{data.frame:
\describe{
  \item{SiteCode}{Site ID code}
  \item{SiteLat}{Latitude}
  \item{SiteLong}{Longitude}
  \item{and so on}{many other variables...}
}}
\usage{
sdat
}
\description{
Site level data
}
\keyword{datasets}
