% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survdat}
\alias{survdat}
\title{Survey level data}
\format{data.frame:
\describe{
  \item{SiteCode}{Site ID code}
  \item{Survey ID}{Survey ID code}
  \item{Depth}{Depth in metres}
  \item{Survey date}{dd/mm/yyyy}
  \item{Visibility}{Observation visibility in metres}
  \item{Diver}{Diver's initials}
}}
\usage{
survdat
}
\description{
Survey level data
}
\keyword{datasets}
