% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgroup_metrics.R
\name{tgroup_metrics}
\alias{tgroup_metrics}
\title{Calculate metrics by trophic groups}
\usage{
tgroup_metrics(datin, ..., tgroups = NA, traits, na.rm = T,
metrics = "biomass", classes = fish_classes())
}
\arguments{
\item{datin}{\code{tbl_df} the main data-frame to use}

\item{...}{optional grouping variables.}

\item{tgroups}{optional \code{character} giving trophic group names}

\item{traits}{\code{data.frame} or  \code{tbl_df} giving trait data}

\item{na.rm}{\code{logical} remove NAs?}

\item{metrics}{\code{character} which metrics to calculate.
see \code{\link{calcmetrics}}}

\item{classes}{\code{character} Filter for Class.}
}
\value{
xout a \code{tbl_df} with the (optional) grouping
variables and the metric value(s) for a trophic group or
set of functional groups
}
\description{
Calculate metrics by trophic groups
}
\details{
see \code{\link{calcmetrics}} for how metrics are calculated
If \code{tgroup = NA} the TrophicGroups field will be ignored and all
trophic groups will be included.
}
\examples{

data(rls)
data (fdat)
flim <- fdat \%>\% select(SpeciesID, Class, TrophicGroup)
herbs <- c("browsing herbivore", "scraping herbivore", "excavator", "algal farmer")
r2 <- rls \%>\% filter(Method ==1) \%>\%
   inner_join(survdat) \%>\%
   inner_join(sdat)
fout <- tgroup_metrics(r2, Location, Ecoregion, tgroups = herbs, traits = flim, na.rm = T, metrics = "alpha")
}
\author{
Christopher J. Brown
}
